feat(#2146): remove the dollar sign from test names
