+++ eo-runtime/src/test/java/EOorg/EOeolang/EObytesEOconcatTest.java
@@ -27,8 +27,6 @@
  */
 package EOorg.EOeolang;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import org.eolang.Data;
 import org.eolang.Dataized;
 import org.eolang.PhMethod;
@@ -36,53 +34,32 @@ import org.eolang.PhWith;
 import org.eolang.Phi;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.api.Test;
 
 /**
- * Test case for {@link EOram$EOram_slice}.
- * {@link EOorg.EOeolang.EOram$EOram_slice} is the generated class. This is the reason
- * why we disable jtcop check.
+ * Test case for {@link EObytes}.
  *
  * @since 0.23
  * @checkstyle TypeNameCheck (4 lines)
- * @checkstyle ParameterNumberCheck (20 lines)
  */
 @SuppressWarnings("JTCOP.RuleAllTestsHaveProductionClass")
-final class EOram$EOram_sliceTest {
+final class EObytesEOconcatTest {
 
-    @ParameterizedTest
-    @CsvSource({
-        "5,  0, hello, 0, 5, hello",
-        "10, 5, hello, 5, 5, hello",
-        "13, 0, hello world, 6, 5, world"
-    })
-    void makesRamSlice(
-        final long total,
-        final int wrt,
-        final String data,
-        final int rdr,
-        final int len,
-        final String result
-    ) throws IOException {
-        final Phi ref = new PhWith(new EOram(Phi.Φ), 0, new Data.ToPhi(total));
-        Ram.INSTANCE.write(ref, wrt, data.getBytes(StandardCharsets.UTF_8));
-        final Phi slice = new PhMethod(ref, "slice");
-        final Phi phi = new PhWith(
+    @Test
+    void concatenatesBytes() {
+        final Phi current = new EOstring$EOas_bytes(new Data.ToPhi("привет "));
+        final Phi provided = new EOstring$EOas_bytes(new Data.ToPhi("mr. ㄤㄠ!"));
+        final Phi phi = new PhMethod(
             new PhWith(
-                slice,
-                "position",
-                new Data.ToPhi((long) rdr)
+                new EObytes$EOconcat(current),
+                "b",
+                provided
             ),
-            "size",
-            new Data.ToPhi((long) len)
+            "as-string"
         );
-        final byte[] bytes = new Dataized(phi).take(byte[].class);
         MatcherAssert.assertThat(
-            new String(bytes, StandardCharsets.UTF_8),
-            Matchers.is(
-                result
-            )
+            new Dataized(phi.copy()).take(String.class),
+            Matchers.equalTo("привет mr. ㄤㄠ!")
         );
     }
