+++ eo-runtime/src/test/java/EOorg/EOeolang/EOramEOram_sliceTest.java
@@ -27,36 +27,63 @@
  */
 package EOorg.EOeolang;
 
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import org.eolang.Data;
 import org.eolang.Dataized;
+import org.eolang.PhMethod;
+import org.eolang.PhWith;
 import org.eolang.Phi;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 /**
- * Test case for {@link EObool$EOnot}.
- * {@link EOorg.EOeolang.EObool$EOnot} is the generated class. This is the reason
+ * Test case for {@link EOram$EOram_slice}.
+ * {@link EOorg.EOeolang.EOram$EOram_slice} is the generated class. This is the reason
  * why we disable jtcop check.
  *
- * @since 0.1
- * @todo #2146:30min Remove RuleAllTestsHaveProductionClass suppressing.
- *  This rule was suppressed because the test class is generated and jtcop can't find it in the
- *  classpath. So, when that feature will be implemented in the jtcop, we have to remove all
- *  the similar suppression's from that test package.
- *  jtcop issue: https://github.com/volodya-lombrozo/jtcop/issues/178
+ * @since 0.23
  * @checkstyle TypeNameCheck (4 lines)
+ * @checkstyle ParameterNumberCheck (20 lines)
  */
 @SuppressWarnings("JTCOP.RuleAllTestsHaveProductionClass")
-final class EObool$EOnotTest {
+final class EOramEOram_sliceTest {
 
-    @Test
-    void inversesValue() {
-        final Phi left = new Data.ToPhi(true);
-        final Phi not = left.attr("not").get();
+    @ParameterizedTest
+    @CsvSource({
+        "5,  0, hello, 0, 5, hello",
+        "10, 5, hello, 5, 5, hello",
+        "13, 0, hello world, 6, 5, world"
+    })
+    void makesRamSlice(
+        final long total,
+        final int wrt,
+        final String data,
+        final int rdr,
+        final int len,
+        final String result
+    ) throws IOException {
+        final Phi ref = new PhWith(new EOram(Phi.Φ), 0, new Data.ToPhi(total));
+        Ram.INSTANCE.write(ref, wrt, data.getBytes(StandardCharsets.UTF_8));
+        final Phi slice = new PhMethod(ref, "slice");
+        final Phi phi = new PhWith(
+            new PhWith(
+                slice,
+                "position",
+                new Data.ToPhi((long) rdr)
+            ),
+            "size",
+            new Data.ToPhi((long) len)
+        );
+        final byte[] bytes = new Dataized(phi).take(byte[].class);
         MatcherAssert.assertThat(
-            new Dataized(not).take(Boolean.class),
-            Matchers.equalTo(false)
+            new String(bytes, StandardCharsets.UTF_8),
+            Matchers.is(
+                result
+            )
         );
     }
+
 }
