+++ eo-runtime/src/test/java/EOorg/EOeolang/EOstringEOas_bytesTest.java
@@ -29,35 +29,28 @@ package EOorg.EOeolang;
 
 import org.eolang.Data;
 import org.eolang.Dataized;
-import org.eolang.PhMethod;
-import org.eolang.PhWith;
 import org.eolang.Phi;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 
 /**
- * Test case for {@link EOint$EOminus}.
- * {@link EOint$EOminus} is the generated class. This is the reason
- * why we disable jtcop check.
+ * Test case for {@link EOstring}.
  *
- * @since 0.1
+ * @since 0.17
  * @checkstyle TypeNameCheck (4 lines)
  */
 @SuppressWarnings("JTCOP.RuleAllTestsHaveProductionClass")
-final class EOint$EOminusTest {
+final class EOstringEOas_bytesTest {
 
     @Test
-    void subtractsNumber() {
-        final Phi left = new Data.ToPhi(42L);
-        final Phi right = new Data.ToPhi(13L);
-        final Phi sub = new PhWith(
-            new PhMethod(left, "minus"),
-            0, right
-        );
+    void convertsStringToBytes() {
+        final Phi str = new Data.ToPhi("Hello, друг!");
+        final Phi phi = new EOstring$EOas_bytes(str);
         MatcherAssert.assertThat(
-            new Dataized(sub).take(Long.class),
-            Matchers.equalTo(29L)
+            new Dataized(phi).take(byte[].class).length,
+            Matchers.equalTo(16)
         );
     }
+
 }
