+++ eo-runtime/src/test/java/EOorg/EOeolang/EOtupleEOatTest.java
@@ -41,7 +41,8 @@ import org.junit.jupiter.api.Test;
  * @since 0.1
  * @checkstyle TypeNameCheck (2 lines)
  */
-final class EOtuple$EOatTest {
+@SuppressWarnings("JTCOP.RuleAllTestsHaveProductionClass")
+final class EOtupleEOatTest {
 
     @Test
     void pushesAndGetsBack() {
